Move arg_from_env, opt_from_env to a utils submodule
* I know this stuff better now and this is something we rarely do
  anymore. So, remove it from the core Command API.
