+++ pydexec/tests/test_command.py
@@ -10,16 +10,7 @@ from testtools.assertions import assert_that
 from testtools.matchers import Equals, Not
 
 from pydexec.command import Command
-from pydexec.tests.helpers import captured_lines
-
-
-def parse_env_output(out_lines):
-    """ Parse the output of the ``env`` command into a dict. """
-    cmd_env = {}
-    for l in out_lines:
-        env_key, env_val = l.split('=', 1)
-        cmd_env[env_key] = env_val
-    return cmd_env
+from pydexec.tests.helpers import captured_lines, parse_env_output
 
 
 def run_cmd(cmd):
@@ -215,163 +206,6 @@ class TestCommand(object):
 
         out_lines, _ = captured_lines(capfd)
         assert_that(out_lines, Equals([]))
-
-    def test_arg_from_env(self, capfd, runner):
-        """
-        When a program argument is specified via an environment variable, the
-        value of the argument should be determined by the environment variable
-        and the variable should be removed from the environment.
-        """
-        cmd = Command('/bin/sh').args('-c', 'echo "$@" && env', '--')
-        cmd.arg_from_env('HOME')  # Pick something that should be in the env
-        runner(cmd)
-
-        out_lines, _ = captured_lines(capfd)
-        assert_that(out_lines.pop(0), Equals(os.environ['HOME']))
-
-        cmd_env = parse_env_output(out_lines)
-        assert_that('HOME' in cmd_env, Equals(False))
-
-    def test_arg_from_env_not_present(self, capfd, runner):
-        """
-        When a program argument is specified via an environment variable, but
-        the variable is not present in the environment, the argument should not
-        be set.
-        """
-        cmd = Command('/bin/sh').args('-c', 'echo "$@" && env', '--')
-        cmd.arg_from_env('DOESNOTEXIST')
-        runner(cmd)
-
-        out_lines, _ = captured_lines(capfd)
-        assert_that(out_lines.pop(0), Equals(''))
-
-        cmd_env = parse_env_output(out_lines)
-        assert_that('DOESNOTEXIST' in cmd_env, Equals(False))
-
-    def test_arg_from_env_no_remove(self, capfd, runner):
-        """
-        When a program argument is specified via an environment variable, and
-        the remove option is set False, the argument should be added and the
-        variable should not be removed from the environment.
-        """
-        cmd = Command('/bin/sh').args('-c', 'echo "$@" && env', '--')
-        cmd.arg_from_env('HOME', remove=False)
-        runner(cmd)
-
-        out_lines, _ = captured_lines(capfd)
-        assert_that(out_lines.pop(0), Equals(os.environ['HOME']))
-
-        cmd_env = parse_env_output(out_lines)
-        assert_that(cmd_env['HOME'], Equals(os.environ['HOME']))
-
-    def test_arg_from_env_default(self, capfd, runner):
-        """
-        When a program argument is specified via an environment variable, and a
-        default value is provided, then if the variable is not set, the default
-        value should be used as the program argument.
-        """
-        cmd = Command('/bin/sh').args('-c', 'echo "$@" && env', '--')
-        cmd.arg_from_env('DOESNOTEXIST', default='foobar')
-        runner(cmd)
-
-        out_lines, _ = captured_lines(capfd)
-        assert_that(out_lines.pop(0), Equals('foobar'))
-
-        # Make extra sure the env variable isn't there
-        cmd_env = parse_env_output(out_lines)
-        assert_that('DOESNOTEXIST' in cmd_env, Equals(False))
-
-    def test_arg_from_env_required(self):
-        """
-        When a program argument is specified via an environment variable, and
-        the required option is set True, then if the variable is not set, an
-        exception should be raised.
-        """
-        with ExpectedException(
-            RuntimeError,
-            'Environment variable "DOESNOTEXIST" is required to determine an '
-                'argument for program "/bin/sh"'):
-            Command('/bin/sh').arg_from_env('DOESNOTEXIST', required=True)
-
-    def test_opt_from_env(self, capfd, runner):
-        """
-        When a program option is specified via an environment variable, the
-        value of the option should be determined by the environment variable
-        and the variable should be removed from the environment.
-        """
-        cmd = Command('/bin/sh').args('-c', 'echo "$@" && env', '--')
-        cmd.opt_from_env('--home', 'HOME')
-        runner(cmd)
-
-        out_lines, _ = captured_lines(capfd)
-        assert_that(out_lines.pop(0),
-                    Equals('--home %s' % (os.environ['HOME']),))
-
-        cmd_env = parse_env_output(out_lines)
-        assert_that('HOME' in cmd_env, Equals(False))
-
-    def test_opt_from_env_not_present(self, capfd, runner):
-        """
-        When a program option is specified via an environment variable, but the
-        variable is not present in the environment, the option should not be
-        set.
-        """
-        cmd = Command('/bin/sh').args('-c', 'echo "$@" && env', '--')
-        cmd.opt_from_env('--home', 'DOESNOTEXIST')
-        runner(cmd)
-
-        out_lines, _ = captured_lines(capfd)
-        assert_that(out_lines.pop(0), Equals(''))
-
-        cmd_env = parse_env_output(out_lines)
-        assert_that('DOESNOTEXIST' in cmd_env, Equals(False))
-
-    def test_opt_from_env_no_remove(self, capfd, runner):
-        """
-        When a program option is specified via an environment variable, and the
-        remove option is set False, the option should be added and the variable
-        should not be removed from the environment.
-        """
-        cmd = Command('/bin/sh').args('-c', 'echo "$@" && env', '--')
-        cmd.opt_from_env('--home', 'HOME', remove=False)
-        runner(cmd)
-
-        out_lines, _ = captured_lines(capfd)
-        assert_that(out_lines.pop(0),
-                    Equals('--home %s' % (os.environ['HOME']),))
-
-        cmd_env = parse_env_output(out_lines)
-        assert_that(cmd_env['HOME'], Equals(os.environ['HOME']))
-
-    def test_opt_from_env_default(self, capfd, runner):
-        """
-        When a program option is specified via an environment variable, and a
-        default value is provided, then if the variable is not set, the default
-        value should be used as the program option.
-        """
-        cmd = Command('/bin/sh').args('-c', 'echo "$@" && env', '--')
-        cmd.opt_from_env('--home', 'DOESNOTEXIST', default='foobar')
-        runner(cmd)
-
-        out_lines, _ = captured_lines(capfd)
-        assert_that(out_lines.pop(0), Equals('--home foobar'))
-
-        # Make extra sure the env variable isn't there
-        cmd_env = parse_env_output(out_lines)
-        assert_that('DOESNOTEXIST' in cmd_env, Equals(False))
-
-    def test_opt_from_env_required(self):
-        """
-        When a program option is specified via an environment variable, and the
-        required option is set True, then if the variable is not set, an
-        exception should be raised.
-        """
-        with ExpectedException(
-            RuntimeError,
-            'Environment variable "DOESNOTEXIST" is required to determine '
-                'option "--home" for program "/bin/sh"'):
-            Command('/bin/sh').opt_from_env('--home', 'DOESNOTEXIST',
-                                            required=True)
 
     def test_workdir_changes_directory(self, capfd, tmpdir, runner):
         """
